tests/wpt/web-platform-tests/html/dom/render-blocking/remove-attr-stylesheet-link-keeps-blocking.tentative.html

<!DOCTYPE html>
<title>Parser-inserted stylesheet link still blocks rendering after removing `blocking=render`</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-render-blocking.js"></script>

<script>
// Test script must be added before the stylesheet link because the stylesheet
// link is script-blocking.

promise_setup(async () => {
  let sheet = await nodeInserted(document.head, node => node.id === 'sheet');
  sheet.blocking = '';
});

test_render_blocking(
  () => {
    let color = getComputedStyle(document.querySelector('.target')).color;
    assert_equals(color, 'rgb(255, 0, 0)');
  },
  'Render-blocking stylesheet is applied');
</script>

<link id="sheet" rel="stylesheet" blocking="render"
      href="support/target-red.css?pipe=trickle(d1)">

<div class="target">Some text</div>
